Healthcare_project/care-connect/Project/healthcare_project/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.61.1
dj-rest-auth==5.0.2
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0 